chore: move JsRequest and JsResponse in separate files
