+++ src/cli/javascript/http_filter.rs
@@ -3,7 +3,8 @@ use std::sync::Arc;
 
 use futures_util::Future;
 
-use super::channel::{JsResponse, Message, MessageContent};
+use super::channel::{Message, MessageContent};
+use super::JsResponse;
 use crate::http::Response;
 use crate::{HttpIO, WorkerIO};
